bring in v2 embedding
