--- example/android/app/.settings/org.eclipse.buildship.core.prefs
@@ -1,2 +1,0 @@
-connection.project.dir=..
-eclipse.preferences.version=1
