+++ example/android/gradle.properties
@@ -1,2 +1,4 @@
 org.gradle.jvmargs=-Xmx1536M
 android.enableR8=true
+android.useAndroidX=true
+android.enableJetifier=true
