+++ pubspec.yaml
@@ -1,17 +1,18 @@
 name: audio_recorder
 description: A Flutter plugin to record audio.
-version: 1.0.2
+version: 1.1.0
 homepage: https://github.com/ZaraclaJ/audio_recorder
 
 dependencies:
-  path: ^1.5.1
+  path: ^1.6.4
   flutter:
     sdk: flutter
-  file: ^5.0.4
+  file: ^5.1.0
 
 
 environment:
-  sdk: ">=2.0.0-dev.40.0.flutter-f981f09760 <3.0.0"
+  sdk: ">=2.0.0-dev.28.0 <3.0.0"
+  flutter: ">=1.12.13+hotfix.6 <2.0.0"
 
 flutter:
   plugin:
@@ -21,5 +22,5 @@ flutter:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  test: ^1.3.0
-  path_provider: ^1.1.0
+  test: ^1.14.4
+  path_provider: ^1.6.10
